Module 2/Week 2/ETL Practice Exercise/employee-activity-dashboard-streamlit/requirements.txt
streamlit>=1.37
oracledb>=2.4
pandas>=2.2
python-dotenv>=1.0
